fix some ptr arithmetic to work in 32 or 64 bit

+++ dcerpc/idl_compiler/sysdep.c
@@ -421,7 +421,7 @@ int output_line
         ** Restore the saved character, update length and pointers, try again.
         */
         *cp = temp;
-        buff_len = buff_len - ((long)cp - (long)buff);
+        buff_len = buff_len - ((uintptr_t)cp - (uintptr_t)buff);
         buff = cp;
     }
